move hermione test types to related directory
