+++ packages/allure-hermione/test/spec/parameters.test.ts
@@ -2,7 +2,7 @@ import { Parameter } from "allure-js-commons";
 import { expect } from "chai";
 import Hermione from "hermione";
 import { beforeEach, describe, it } from "mocha";
-import { HermioneAllure } from "../../src";
+import { HermioneAllure } from "../types";
 
 describe("parameters", () => {
   let hermione: HermioneAllure;
