+++ packages/allure-hermione/test/spec/steps.test.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import Hermione from "hermione";
 import { beforeEach, describe, it } from "mocha";
 import Sinon from "sinon";
-import { HermioneAllure } from "../../src";
+import { HermioneAllure } from "../types";
 
 describe("steps", () => {
   let hermione: HermioneAllure;
